requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Migrate
Flask-Cors
Flask-smorest
python-dotenv
marshmallow
SQLAlchemy
gunicorn
psycopg2-binary
pytest
Flask-Limiter
redis
Flask-Caching
bleach
textblob
nltk